requirements.txt
opencv-python==4.8.0.74
pytesseract==0.3.10
Pillow==9.4.0
numpy==1.24.2